requirements.txt
openai==1.65.5
websockets==13.1
loguru==0.7.3
python-dotenv==1.0.1
requests==2.32.3
flask==2.3.3
flask-socketio==5.3.6
psutil==5.9.6
werkzeug==2.3.7
pytz==2023.3
aiosqlite==0.19.0
pydantic==2.5.0
